pyproject.toml
[project]
name = "classy-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=6.0",
    "django-crispy-forms>=2.5",
    "django-dynamic-preferences>=1.17.0",
    "django-extra-views>=0.16.0",
    "django-filter>=25.2",
    "django-htmx>=1.27.0",
    "django-render-block>=0.11",
    "django-select2>=8.4.3",
    "django-simple-menu>=2.1.4",
    "django-tables2>=2.8.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-stubs>=5.2.8",
    "djlint>=1.36.4",
    "flake8>=7.3.0",
    "flake8-isort>=7.0.0",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.8",
]
